requirements.txt
streamlit
google-generativeai
faiss-cpu
python-dotenv
langchain
PyPDF2
langchain_google_genai
chromadb
pdfminer.six
fitz